move StateConverter to worker, so that server and worker can both access without needing to introduce any new dependencies
