+++ airbyte-server/src/main/java/io/airbyte/server/handlers/StateHandler.java
@@ -8,8 +8,8 @@ import io.airbyte.api.model.generated.ConnectionIdRequestBody;
 import io.airbyte.api.model.generated.ConnectionState;
 import io.airbyte.api.model.generated.ConnectionStateCreateOrUpdate;
 import io.airbyte.config.StateWrapper;
-import io.airbyte.config.helpers.StateConverter;
 import io.airbyte.config.persistence.StatePersistence;
+import io.airbyte.workers.helper.StateConverter;
 import java.io.IOException;
 import java.util.Optional;
 import java.util.UUID;
