+++ airbyte-server/src/test/java/io/airbyte/server/handlers/StateHandlerTest.java
@@ -18,13 +18,13 @@ import io.airbyte.commons.enums.Enums;
 import io.airbyte.commons.json.Jsons;
 import io.airbyte.config.StateType;
 import io.airbyte.config.StateWrapper;
-import io.airbyte.config.helpers.ProtocolConverters;
 import io.airbyte.config.persistence.StatePersistence;
 import io.airbyte.protocol.models.AirbyteGlobalState;
 import io.airbyte.protocol.models.AirbyteStateMessage;
 import io.airbyte.protocol.models.AirbyteStateMessage.AirbyteStateType;
 import io.airbyte.protocol.models.AirbyteStreamState;
 import io.airbyte.protocol.models.StreamDescriptor;
+import io.airbyte.workers.helper.ProtocolConverters;
 import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
