+++ airbyte-workers/src/main/java/io/airbyte/workers/helper/StateConverter.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2022 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.config.helpers;
+package io.airbyte.workers.helper;
 
 import io.airbyte.api.model.generated.ConnectionState;
 import io.airbyte.api.model.generated.ConnectionStateType;
